+++ CHANGELOG.md
@@ -4,7 +4,10 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 
 #### Features supported with current release:
 - Deployment plan
+- Interconnect
 - Plan Scripts
+- Storage pool
+- Storage volume
 
 # 4.5.0
 #### Notes
@@ -28,6 +31,7 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 - Logical switch group
 - Managed SAN
 - OS Build Plan
+- SAS interconnect
 - SAS interconnect type
 - SAS logical interconnect
 - SAS logical interconnect group